+++ src/modules/canvas/hooks/index.ts
@@ -1,2 +1,4 @@
 export { useNotesAsNodes } from './useNotesAsNodes';
 export { useNodePositionSync } from './useNodePositionSync';
+export { useConnectionsAsEdges } from './useConnectionsAsEdges';
+export { useConnectionHandlers } from './useConnectionHandlers';